modify: sellingView's design
